% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_modis.R
\name{add_raster_attributes}
\alias{add_raster_attributes}
\title{Add attributes to a Raster* object}
\usage{
add_raster_attributes(r,attribute)
}
\arguments{
\item{r}{A \code{\link[raster]{Raster-class}} object to add attributes (or categories) to}

\item{attribute}{a character \code{vector} of class names, with the same length as the number
of unique numeric values in the input raster object}
}
\description{
Convert a Raster* object with numeric values to classes
}
\details{
The function uses rgdal::writeGDAL with \code{catNames} argument to write a tif file to disk with attributes
and read the tif back to R
}
\author{
Antoine Stevens
}

