% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_modis.R
\name{convert_dn_modis}
\alias{convert_dn_modis}
\title{Convert MOD13A2 and MOD15A2 DN values to physical values}
\usage{
convert_dn_modis(r,type)
}
\arguments{
\item{r}{A \code{\link[raster]{Raster-class}} object to convert}

\item{type}{character \code{vector} of length 1 giving the type of MODIS data. Should be one of these values:
'Fpar_1km','Lai_1km','FparLai_QC','FparExtra_QC','FparStdDev_1km','LaiStdDev_1km','NDVI','EVI','VI','red','NIR','view_zenith_angle','sun_zenith_angle','relative_azimuth_angle','composite_day_of_the_year','pixel_reliability'}
}
\value{
A \code{\link[raster]{Raster-class}} object with converted DN values
}
\description{
Set values outside valid range to \code{NA} and scale input data using pre-determined gain values
}
\note{
It is used in \code{\link{convert_modis}}. This function is obviously valid for MOD13A2 and MOD15A2 only !
}
\author{
Antoine Stevens
}
\references{
Solano et al. 2010. MODIS Vegetation Index User's Guide (MOD13 Series), 42 p.
http://gis.stackexchange.com/questions/84058/reprojected-modis-ndvi-has-range-from-32768-to-32767-expected-1-to-1
}
\seealso{
\code{\link{convert_modis}}
}

