% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_modis.R
\name{convert_modis}
\alias{convert_modis}
\title{Convert MODIS files}
\usage{
convert_modis(path = ".", pattern = NULL, type, convertDN = TRUE,
                     extractAll = FALSE, brick = TRUE, filename, overwrite = TRUE, ...)
}
\arguments{
\item{path}{Path to the folder where MODIS files are stored. Default is the working directory. See \code{\link[MODIS]{preStack}}.}

\item{pattern}{optional regular expression, only file names matching the regular expression will be extracted. See \code{\link[MODIS]{preStack}}.}

\item{type}{character \code{vector} of length 1 giving the type of MODIS data. Should be one of these values:
'Fpar_1km','Lai_1km','FparLai_QC','FparExtra_QC','FparStdDev_1km','LaiStdDev_1km','NDVI','EVI','VI','red','NIR','view_zenith_angle','sun_zenith_angle','relative_azimuth_angle','composite_day_of_the_year','pixel_reliability'}

\item{convertDN}{a logical value indicating whether MODIS DN values should be converted to physical values.
See \code{\link{convert_dn_modis}}. Default is \code{TRUE}.}

\item{extractAll}{a logical value indicating whether time info, file names and date object should be returned in addition to the processed raster object.
Default is \code{FALSE}.}

\item{brick}{logical value indicating whether a \code{\link[raster]{RasterBrick-class}} should be returned. Default is \code{TRUE}.}

\item{filename}{Passed to \code{\link[raster]{writeRaster}}, if \code{extractAll} is \code{FALSE}.}

\item{...}{arguments passed to \code{\link{orgTime}} (for instance to restrict the studied period with \code{begin} or \code{end}
       or \code{\link[raster]{writeRaster}}
or define new time stamps to interpolate to with \code{nDays})}

\item{overwrite}{Passed to \code{\link[raster]{writeRaster}}, if \code{extractAll} is \code{FALSE}.}
}
\value{
a \code{list} with the following elements (where XXX corresponds to the \code{type} argument):
\itemize{
   \item{\code{raster_XXX}}{the processed Raster* object}
   \item{\code{raster_fill_XXX}}{if \code{type} is 'Fpar_1km|Lai_1km|FparStdDev_1km|LaiStdDev_1km', fill values are automatically extracted (https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod15a2)}
   \item{\code{ti_XXX}}{time info as returned by \code{\link[MODIS]{orgTime}}}
   \item{\code{fn_XXX}}{file names as returned by \code{\link[MODIS]{preStack}}}
   \item{\code{d_XXX}}{\code{vector} of dates as returned by \code{\link[MODIS]{extractDate}}}
}
If there is only one element in the \code{list}, the \code{convert_modis} returns the \code{unlist}-ed object.
}
\description{
Convert MODIS files into a Raster* object, and possibly convert raw DN to physical values.
}
\examples{
\dontrun{
# Get and extract NDVI MOD13A2 for JUNE 2009
runGdal(product="MOD13A2", begin = "2009.06.01", end = "2009.06.30",
       tileH = 19, tileV = 5 ,
       SDSstring =  "100000000000",job="H19V5")
# process and convert to raster
path <- "~/COPERNICUS/MODIS_DATA/PROCESSED/H19V5"
convert_modis(path = path,pattern = "16_days_NDVI")
}
}
\author{
Antoine Stevens
}

