% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_qf_modis.R
\name{convert_qf_modis}
\alias{convert_qf_modis}
\title{Convert bit values from a QFLAG MODIS product}
\usage{
convert_qf_modis(r,qf,type,cl = NULL,filename = rasterTmpFile(), ...)
}
\arguments{
\item{r}{A \code{\link[raster]{Raster-class}} object}

\item{qf}{Quality Flag to be extracted. Can be more than one of these:
If type = 'MOD13': 'MODLAND_QA','VI_usefulness','aerosol_quantity','adjacent_cloud','atmosphere_brdf','mixed_clouds','land_water_flag','snow','shadow'.
If type = 'MOD15':'MODLAND_QC','sensor','dead_detector','cloud','scf_qc'.
If type = 'MOD15Extra':'sea', 'snow', 'aerosol', 'cirrus', 'cloud', 'shadow', 'biome'}

\item{type}{character \code{vector} of length 1 giving the MODIS data type. Should be one of these: 'MOD13', 'MOD15', 'MOD15Extra'}

\item{cl}{cluster object for parallel processing. Default is \code{NULL}}

\item{filename}{Passed to \code{\link[raster]{writeRaster}}.}

\item{...}{arguments passed to \code{\link[raster]{writeRaster}}.}
}
\value{
A \code{\link[raster]{RasterBrick-class}} object with two values (1,0), representing respectively pixels that are flagged by at least one of the give \code{qf},
       and pixels that are not flagged in any of the given \code{qf}
}
\description{
Convert a Raster* object representing MODIS QFLAG bit values to a Raster* with two categories representing flagged and non-flagged pixels
}
\author{
Antoine Stevens
}
\references{
https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13a2,
https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod15a2
}
\seealso{
\code{\link[MODIS]{extractBits}} and \code{\link{interpolate_raster}} for an example
}

