% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filesUrl.R
\name{filesUrl}
\alias{filesUrl}
\title{Get MODIS file URL's}
\usage{
filesUrl(url)
}
\arguments{
\item{url}{an url pointing to MODIS data for a given date}
}
\value{
a character \code{vector} of urls pointing to data locations
}
\description{
Get MODIS file URL's
}
\note{
This produce the same results as \code{MODIS:::filesUrl} but uses the \code{rvest}
package instead of \code{RCurl}
}
\examples{
\dontrun{
filesUrl(url = "http://e4ftl01.cr.usgs.gov/MOLT/MOD13A2.005/2015.05.25")
}
}
\seealso{
\code{MODIS:::filesUrl} in the MODIS package
}

