% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_raster.R
\name{gapfill_raster}
\alias{gapfill_raster}
\title{Fill NA values in Raster* time series by interpolation (gap filling)}
\usage{
gapfill_raster(x, w=NULL, t=NULL, timeInfo = orgTime(x), method,
                      lambda = 5000, nIter= 3, outlierThreshold = NULL,
                      df = 6, ws = 2, cl = NULL, filename=rasterTmpFile(),...)
}
\arguments{
\item{x}{A \code{\link[raster]{Raster-class}} object to interpolate}

\item{w}{Optional, a \code{\link[raster]{Raster-class}} object with weight information. See e.g. \code{\link[MODIS]{makeWeights}}.}

\item{t}{Optional, a \code{\link[raster]{Raster-class}} object with MODIS 'composite day of the year' data.}

\item{timeInfo}{object returned by \code{orgTime} defining the input and output time stamps to interpolate to.}

\item{method}{character \code{vector} of length 1 with one of the following value:
'spline' for smoothing splines ,'linear' for linear approximation ,'constant' for constant approximation,
'nn' for nearest neighbour , 'whittaker' for whittaked smoothing.}

\item{lambda}{Yearly lambda value. Default is 5000. See \code{\link[MODIS]{whittaker.raster}}.}

\item{nIter}{numeric \code{vector} of length 1 giving the number of iterations for the upper envelope fitting ('whittaker' \code{method}).
Default is 3. See \code{\link[MODIS]{whittaker.raster}}.}

\item{outlierThreshold}{Numeric. If provided, \code{outlierTreshold} allows to remove outliers before smoothing by fitting a
whittaker function and removing values with residuals higher than \code{outlierTreshold}.}

\item{df}{Yearly degree of freedom for 'spline' \code{method}. Default is 6. See \code{\link[MODIS]{smooth.spline.raster}}}

\item{ws}{window size for the `fensholt` \code{method}. Default to 2. See details}

\item{cl}{cluster object for parallel processing (do not use it if beginCluster() has been called. Default is \code{NULL}}

\item{filename}{name of the file where the interpolated raster is saved. Default is created through \code{\link[raster]{rasterTmpFile}}}

\item{...}{arguments passed to \code{\link[raster]{writeRaster}}}
}
\value{
A \code{\link[raster]{Raster-class}} object with NA values filled
}
\description{
Fill \code{NA} values in Raster* time series
}
\details{
The 'fensholt' \code{method} fill gaps in a time serie according to the following rule:
If t0 is flagged as NA (e.g. cloudy) an average between t-1, t+1 is performed;
if only one of t-1 or t+1 is not missing t0 will be replaced by this value.
If both t-1 and t+1 are missing then the filter is extended to cover also t-2 and t+2 following the same rules as for
t-1 and t+1
}
\note{
Most of the code is borrowed from \code{\link[MODIS]{smooth.spline.raster}} and \code{\link[MODIS]{whittaker.raster}}
}
\examples{
# See interpolate_raster
}
\author{
Antoine Stevens & Matteo Mattiuzzi
}
\references{
Fensholt, R., Rassmussen, K., Nielsen, T.T., and Mbow, C. (2009). Evaluation of earth observation based long term vegetation trends - Intercomparing NDVI time series trend analysis consistency of Sahel from AVHRR GIMMS, Terra MODIS and SPOT VGT data. Remote Sensing of Environment 113, 1886-1898.
}
\seealso{
\code{interpolate_raster}
}

