% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHdf.R
\name{getHdf}
\alias{getHdf}
\title{Create or update a local, user-defined subset of the global online MODIS grid data pool}
\usage{
getHdf(product, begin=NULL, end=NULL, tileH=NULL, tileV=NULL, extent=NULL,
collection=NULL, HdfName, quiet=FALSE, wait=0.5, checkIntegrity=FALSE, ...)
}
\arguments{
\item{product}{Character. (Ignored if \code{HdfName} is provided.) MODIS grid product to be downloaded. For 'MODIS TERRA Vegetation Indices at 250 m ground resolution' use: \code{MOD13Q1}. For vegetation indices from TERRA and AQUA use \code{M.D13Q1}. Run \code{getProduct()} for an overview of supported products.}

\item{begin}{Character. (Ignored if \code{HdfName} is provided.) Subset begin date of MODIS time series. See: \code{?transDate} for formatting.}

\item{end}{Character. (Ignored if \code{HdfName} is provided.) Subset end date of MODIS time series. See: \code{?transDate} for formatting. Can even be a future date, for continuous update in schedule jobs.}

\item{tileH}{Character vector. (Ignored if \code{HdfName} is provided.) Horizontal tile number(s) (for MODIS between 0 and 35). I.e.: \code{tileH = 1:5}. Refer to \url{http://modis-land.gsfc.nasa.gov/MODLAND_grid.htm} to find your area.}

\item{tileV}{Character vector. (Ignored if \code{HdfName} is provided.) Vertical tile number(s) (for MODIS between 0 and 17). I.e.: \code{tileV = 9:17}.}

\item{extent}{see \code{'extent'} in \code{?getTile}.}

\item{collection}{Character or integer. (Ignored if \code{HdfName} is provided.) Refers to the MODIS product collection-version. Often '005', but see MODIS pages or run \code{?getCollection} for more information.}

\item{HdfName}{Character (vector or list). Full HDF file name(s) to download. If just few single (AND known!) files should be downloaded use \code{HdfName}. If \code{HdfName} is provided the other file-related parameters are ignored: \code{begin,end,collection,tileV,...}. If many files have to be checked and downloaded it is simpler, safer and faster to use the following parameters.}

\item{quiet}{Logical. Refers to all internal \code{download.file()} actions. Default \code{FALSE}, if \code{TRUE} the \code{getHdf()} becomes really boring.}

\item{wait}{Numeric. Inserts a break (in seconds) after every internal \code{download.file()} or \code{getURL()} call. It reduces the chance of FTP connection error that often occurs after many requests. (Hopefully a temporary solution!)}

\item{checkIntegrity}{Logical. Default \code{FALSE}, no file integrity check is performed. If \code{TRUE} the HDF files size is checked and the function tries to re-download the file. Currently it uses GDAL (gdalinfo), that supports HDF4, run \code{MODISoptions()} to check availability).}

\item{...}{Arguments found in MODISoptions, sections 'STORAGE' and 'DOWNLOAD'}
}
\value{
See \code{\link[MODIS]{getHdf}} in the MODIS package
}
\description{
See \code{\link[MODIS]{getHdf}} in the \code{MODIS} package, which this function is essentially a copy-paste. The only changes are related
to the way MODIS files are downloaded which uses now the \code{rvest} package.
}
\examples{
\dontrun{
# Get NDVI MOD13A2 for June 2009
getHdf(product="MOD13A2", begin = "2009.06.01", end = "2009.06.30",
       tileH = 19, tileV = 5 ,job="H19V5")
}
}
\author{
Matteo Mattiuzzi and Antoine Stevens
}

