% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgTime.R
\name{orgTime}
\alias{orgTime}
\title{Handles the in/output date used in the filtering/interpolation}
\usage{
orgTime(files, nDays = "asIn", begin = NULL, end = NULL, pillow = 75,
  pos1 = 10, pos2 = 16, format = "\%Y\%j")
}
\arguments{
\item{files}{MODIS filenames, eg output of \code{runGdal}}

\item{nDays}{Integer, \code{Date} vector or character string indicating the time interval/time stamps for the output layers.
Default is "asIn" that includes the exact input dates within the period selected using \code{begin} and \code{end}.
Can also be nDays="1 month" or "1 week"}

\item{begin}{Default is from earliest input dataset. Here you can supply the begin date of the output}

\item{end}{Default to the end of the input dataset. Here you can specify the end date of the output
(note, the exact end-date depends on \code{begin} and \code{nDays} argument.}

\item{pillow}{Number of days added on the beginning and on the end of a time serie.}

\item{pos1}{Start position of date in the filename.}

\item{pos2}{End position of date in the filename.}

\item{format}{How is the date formatted in the file, default expects: 'YYYYDDD' ("\%Y\%j").}
}
\value{
See \code{\link[MODIS]{orgTime}} in the MODIS package
}
\description{
See \code{\link[MODIS]{orgTime}} in the MODIS package, which this function is a copy-paste, with one modification
             of the \code{nDays} argument which accepts now user-defined time series
}
\author{
Matteo Mattiuzzi and Antoine Stevens
}

