% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGdal.R
\name{runGdal}
\alias{runGdal}
\title{Process MODIS hdf with GDAL}
\usage{
runGdal(product, collection=NULL, begin=NULL, end=NULL, extent=NULL, tileH=NULL,
tileV=NULL, buffer=0, SDSstring=NULL, job=NULL, checkIntegrity=TRUE, wait=0.5,
quiet=FALSE,gdalPath = "",...)
}
\arguments{
\item{product}{See: \code{?getProduct}}

\item{collection}{Default is to download most recent collection version. See: \code{?getCollection}}

\item{begin}{Default is from the beginning of data. See: \code{?transDate}}

\item{end}{Default is to the end of data. See: \code{?transDate}}

\item{extent}{Default is NULL for interactive selection. See: \code{?getTile} and Details!}

\item{tileH}{Numeric. Default is NULL for interactive selection. See: \code{?getTile}}

\item{tileV}{Numeric. Default is NULL for interactive selection. See: \code{?getTile}}

\item{buffer}{Numeric. Buffer [in units of the 'outProj'] around the specified extent. See: \code{?getTile}}

\item{SDSstring}{Default is extract all SDS (layers). See: \code{?getSds}.}

\item{job}{Character name. Name of the current job for the creation of the output folder. If not specified it is created: 'PRODUCT.COLLECTION_DATETIME'.}

\item{checkIntegrity}{Default \code{TRUE}. Check if the file is 'healthy' (and download it again if not) before doing further processings with it. This consumes also a littel bit of time to do the check.}

\item{wait}{Numeric. Time to wait to re-try a un-successful online connection (URL listenting or 'download.file()' process). Keep it small, max ~1 sec.}

\item{quiet}{Logical, Default \code{FALSE}. Some progress informations.}

\item{gdalPath}{Path to the gdal binaries.}

\item{...}{See section 'Methods'.}
}
\value{
See \code{\link[MODIS]{runGdal}} in the MODIS package
}
\description{
Download MODIS data from archive and process the files with GDAL. This function
is basically copy-pasted from MODIS::runGdal, with a few bugs corrected. It uses also now the \code{GdalUtils}
package, to simplify the code
}
\examples{
\dontrun{
# Get and extract NDVI MOD13A2 for June 2009
runGdal(product="MOD13A2", begin = "2009.06.01", end = "2009.06.30",
       tileH = 19, tileV = 5 ,
       SDSstring =  "100000000000",job="H19V5",
       gdalPath = "C:/OSGeo4W64/bin")
}

}
\author{
Matteo Mattiuzzi and Antoine Stevens
}

